添加首页搜索

+++ src/main/java/com/example/blog/reporsitory/UserRepository.java
@@ -4,8 +4,9 @@ import com.example.blog.entity.User;
 import org.springframework.data.repository.CrudRepository;
 
 
-public interface UserRepository extends CrudRepository<User,Long> {
+public interface UserRepository extends CrudRepository<User,Integer> {
     User findByUserEmailAndUserPassword(String userEmail,String userPassword);
     User findByUserId(Integer userId);
+    //void deleteByUserId(Integer userId);
 
 }
